+++ src/helpers/pdf.ts
@@ -1,0 +1,5 @@
+import { Novel } from "../models/novel/Novel";
+
+export const GeneratePDF = (novel: Novel) => {
+  console.log(novel);
+};