+++ src/models/release/index.ts
@@ -1,11 +1,8 @@
 /* tslint:disable:no-console */
 
-import octokit, {
-  ReposGetLatestReleaseResponseAssetsItem,
-} from "@octokit/rest";
+import octokit from "@octokit/rest";
 import fs from "fs";
 import moment = require("moment");
-import os from "os";
 import path from "path";
 import request = require("request");
 import progress = require("request-progress");
@@ -13,45 +10,23 @@ import { log } from "winston";
 
 import { WrapTMCT } from "../../apis/loggerWrapper";
 import { COLORS } from "../../constants/color.const";
+import { GIT_REPONAME, GIT_USERNAME } from "../../constants/config.const";
 import { DIR_TMP_RANDOM } from "../../constants/file.const";
 import { ND } from "../../constants/nd.const";
 import { FormatFileSize } from "../../helpers/helper";
 
-const owner = "kamontat";
-const repo = "nd-js";
+import { GetOSName } from "./helper";
+import { InstallOption, VersionObject } from "./model";
 
-export interface InstallOption {
-  bin: string;
-}
-
-export interface VersionObject {
-  version: string;
-  assets: Array<ReposGetLatestReleaseResponseAssetsItem>;
-  url: string;
-  date: string;
-}
-
-const GetOSName = () => {
-  switch (os.platform()) {
-    case "win32":
-      return "win.exe";
-    case "linux":
-      return "linux";
-    case "darwin":
-      return "macos";
-    default:
-      return undefined;
-  }
-};
+const github = new octokit();
 
 export const GetLatestVersion = async () => {
-  const github = new octokit();
   github.authenticate({
     type: "token",
     token: process.env.GITHUB_TOKEN || "",
   });
 
-  const result = await github.repos.getLatestRelease({ owner, repo });
+  const result = await github.repos.getLatestRelease({ owner: GIT_USERNAME, repo: GIT_REPONAME });
 
   return {
     version: result.data.tag_name,
@@ -65,18 +40,15 @@ export const ListAllVersion = async () => {
   const github = new octokit();
 
   const result = await github.repos.listReleases({
-    owner,
-    repo,
+    owner: GIT_USERNAME,
+    repo: GIT_REPONAME,
     per_page: 100,
   });
 
   return result.data;
 };
 
-export const InstallSpecifyVersion = (
-  version: string,
-  opts?: InstallOption,
-) => {
+export const InstallSpecifyVersion = (version: string, opts?: InstallOption) => {
   return ListAllVersion().then(versions => {
     const specify = versions.find(v => v.tag_name === version);
     if (specify)
@@ -92,10 +64,7 @@ export const InstallSpecifyVersion = (
   });
 };
 
-export const InstallVersion = (
-  version: VersionObject,
-  opts?: InstallOption,
-) => {
+export const InstallVersion = (version: VersionObject, opts?: InstallOption) => {
   const defaultOptions = {
     bin: "/usr/local/bin",
   } as InstallOption;
@@ -111,10 +80,7 @@ export const InstallVersion = (
     return;
   }
 
-  const asset = version.assets.find(
-    v => !v.name.includes("admin") && v.name.includes(osName || "undefined"),
-  );
-
+  const asset = version.assets.find(v => !v.name.includes("admin") && v.name.includes(osName || "undefined"));
   if (!asset) {
     log(WrapTMCT("error", "Error", "Binary file not found"));
     return;
@@ -130,21 +96,13 @@ export const InstallVersion = (
     WrapTMCT(
       "info",
       "Installation",
-      `Version ${COLORS.Important.color(version.version)} ${COLORS.Dim.color(
-        version.date,
-      )}`,
+      `Version ${COLORS.Important.color(version.version)} ${COLORS.Dim.color(version.date)}`,
     ),
   );
 
   const size = FormatFileSize(asset.size);
 
-  log(
-    WrapTMCT(
-      "info",
-      "Binary",
-      `Downloading ${COLORS.Name.color(name)} size=${size}`,
-    ),
-  );
+  log(WrapTMCT("info", "Binary", `Downloading ${COLORS.Name.color(name)} size=${size}`));
 
   progress(request(downloadURL))
     .on("progress", (state: any) => {
@@ -161,12 +119,8 @@ export const InstallVersion = (
       const completed = "#";
       const incompleted = "-";
 
-      const header = `${FormatFileSize(
-        state.size.transferred,
-      )}/${FormatFileSize(state.size.total)}`.padStart(title);
-      const progressbar = `[${""
-        .padStart(block, completed)
-        .padEnd(barSize, incompleted)}]`;
+      const header = `${FormatFileSize(state.size.transferred)}/${FormatFileSize(state.size.total)}`.padStart(title);
+      const progressbar = `[${"".padStart(block, completed).padEnd(barSize, incompleted)}]`;
 
       const speed = `${FormatFileSize(state.speed)}/s`;
       const percentString = `${(percent * 100).toFixed(2)}%`.padEnd(3);
@@ -174,21 +128,13 @@ export const InstallVersion = (
       moment.locale("en");
       const time = moment.duration(state.time.remaining, "seconds");
 
-      const footer = `${percentString}  ${speed}  ${time.humanize()}`.padEnd(
-        footerSize,
-      );
+      const footer = `${percentString}  ${speed}  ${time.humanize()}`.padEnd(footerSize);
 
       process.stdout.write(`${header} ${progressbar} ${footer} \x1b[0G`);
     })
     .on("end", () => {
       console.log("\n");
-      log(
-        WrapTMCT(
-          "info",
-          "Result",
-          `${COLORS.Name.color(name)} was saved on ${dest}`,
-        ),
-      );
+      log(WrapTMCT("info", "Result", `${COLORS.Name.color(name)} was saved on ${dest}`));
 
       fs.renameSync(dest, path.join(options.bin, ND.PROJECT_NAME));
     })